fix(lyrics): split deterministic tests into sibling file (1000-line cap)
claude_merge_tests.rs hit 1004 lines after adding 7 merge_deterministic
tests, breaking the airuleset 1000-line file cap. Moved the new tests +
their helpers into a new sibling `claude_merge_deterministic_tests.rs`
included via a second `#[cfg(test)] #[path = ...] mod` declaration in
claude_merge.rs. No behavior change.
